requirements.txt
scikit-learn
numpy
streamlit==1.31.0
matplotlib
pandas
imageio
openpyxl
plotly
streamlit-geolocation
sqlalchemy
SQLAlchemy-Utils
python-dotenv==1.0.0
psycopg2-binary==2.9.9